# iCanteen-api接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "iCanteen-api接口",
    "description": "校园“智汇”食堂后端接口文档（返回统一使用 Result 包装）",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/admin/user/list": {
      "get": {
        "summary": "管理员分页查询用户列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理员"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": false,
            "example": 0,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "example": 0,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "phone",
            "in": "query",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nickname",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=AdminUserDTO[], total=count)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/user/{id}": {
      "get": {
        "summary": "管理员查看用户详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理员"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1070,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=AdminUserDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "管理员修改用户详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理员"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1070,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdateDTO"
              },
              "example": {
                "birthday": "2001-06-10",
                "city": "Beijing",
                "gender": false,
                "phone": "13456789005",
                "icon": "/imgs/icons/default3.png",
                "introduce": "hello!",
                "nickname": "testuser_03",
                "password": "12345678",
                "credits": 1,
                "role": 0
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/food-order/{id}": {
      "get": {
        "summary": "用户查询订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2595113729524213,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderDetailDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/food-order/list": {
      "get": {
        "summary": "管理员分页查询订单",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单-管理"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1,
                2,
                3,
                4
              ]
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "windowId",
            "in": "query",
            "description": "",
            "required": false,
            "example": "4",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderSimpleDTO[],total=总数)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/food-order/{id}": {
      "get": {
        "summary": "管理员查询订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单-管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2342921839837993,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderDetailDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/food-order/{id}/status": {
      "put": {
        "summary": "管理员更新订单状态",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单-管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2342921839837993,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodOrderStatusUpdateDTO"
              },
              "example": {
                "status": 1,
                "cancelReason": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMsg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "null"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
...
      "AdminUserUpdateDTO": {
        "type": "object",
        "description": "管理员修改用户详情参数",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "nickname": {
            "type": "string",
            "maxLength": 32
          },
          "icon": {
            "type": "string",
            "maxLength": 255
          },
          "city": {
            "type": "string",
            "maxLength": 64
          },
          "introduce": {
            "type": "string",
            "maxLength": 256
          },
          "gender": {
            "type": "boolean",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "credits": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "FoodOrderStatusUpdateDTO": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "description": "0待接单,1制作中,2待取餐,3已完成,4已取消",
            "enum": [
              0,
              1,
              2,
              3,
              4
            ]
          },
          "cancelReason": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}